change ID type from Integer to long

+++ JdbcTemplate/src/main/java/com/farzadafi/jdbctemplate/repository/UniversityRepository.java
@@ -3,5 +3,5 @@ package com.farzadafi.jdbctemplate.repository;
 import com.farzadafi.jdbctemplate.base.repository.BaseRepository;
 import com.farzadafi.jdbctemplate.entity.University;
 
-public interface UniversityRepository extends BaseRepository<Integer, University> {
+public interface UniversityRepository extends BaseRepository<Long, University> {
 }
